Davide/aule/requirements.txt
flask
flask-restful
google-cloud-firestore
gunicorn
wtforms
